requirements.txt
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
Flask==3.0.2
Flask-WTF==1.2.1
WTForms==3.1.2 